done room stats, done changing room, done boss room, done items room

+++ godot_game_STUDIO_project/godot_game_STUDIO_project/Enemies.cpp
@@ -11,7 +11,7 @@ godot::Enemies::Enemies()
 
 int godot::Enemies::_get_enemies_count()
 {
-	return enemies.size();
+	return (int)enemies.size();
 }
 
 void godot::Enemies::_add_enemy(Node2D* enemy)
